+++ src/main/java/application/repository/UserRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface UserRepository extends JpaRepository<Account, Long> {
 	 Account findByEmail(String email);
-	 Account findById(int id);
+	 Account findById(Long id);
 }
